move code into __init__.py - cleaner module import
